[REFACTOR] seprate test in category and refactor smell code
